BaseAnyHolder as proc-macro
