+++ ruststep/tests/any_deserialize.rs
@@ -199,82 +199,11 @@ impl WithVisitor for Sub2Holder {
     }
 }
 
-#[derive(Clone, Debug, PartialEq)]
+#[derive(Clone, Debug, PartialEq, ruststep_derive::Holder)]
+#[holder(table = Table, field = base)]
 enum BaseAny {
     Sub1(Box<Sub1>),
     Sub2(Box<Sub2>),
-}
-
-#[derive(Clone, Debug, PartialEq)]
-enum BaseAnyHolder {
-    Sub1(Box<Sub1Holder>),
-    Sub2(Box<Sub2Holder>),
-}
-
-impl Holder for BaseAnyHolder {
-    type Owned = BaseAny;
-    type Table = Table;
-    fn into_owned(self, table: &Table) -> Result<Self::Owned> {
-        Ok(match self {
-            BaseAnyHolder::Sub1(sub) => BaseAny::Sub1(Box::new(sub.into_owned(table)?)),
-            BaseAnyHolder::Sub2(sub) => BaseAny::Sub2(Box::new(sub.into_owned(table)?)),
-        })
-    }
-    fn name() -> &'static str {
-        "BaseAny"
-    }
-    fn attr_len() -> usize {
-        0
-    }
-}
-
-impl<'de> de::Deserialize<'de> for BaseAnyHolder {
-    fn deserialize<D>(deserializer: D) -> ::std::result::Result<Self, D::Error>
-    where
-        D: de::Deserializer<'de>,
-    {
-        deserializer.deserialize_tuple_struct("A", 2, BaseAnyHolderVisitor {})
-    }
-}
-
-struct BaseAnyHolderVisitor;
-
-impl<'de> ::serde::de::Visitor<'de> for BaseAnyHolderVisitor {
-    type Value = BaseAnyHolder;
-    fn expecting(&self, formatter: &mut ::std::fmt::Formatter) -> ::std::fmt::Result {
-        write!(formatter, "BaseAny")
-    }
-
-    // Entry point for Record or Parameter::Typed
-    fn visit_map<A>(self, mut map: A) -> ::std::result::Result<Self::Value, A::Error>
-    where
-        A: ::serde::de::MapAccess<'de>,
-    {
-        let key: String = map
-            .next_key()?
-            .expect("Empty map cannot be accepted as ruststep Holder"); // this must be a bug, not runtime error
-        match key.as_str() {
-            "SUB1" => {
-                let value: Sub1Holder = map.next_value()?; // send to Self::visit_seq
-                return Ok(BaseAnyHolder::Sub1(Box::new(value)));
-            }
-            "SUB2" => {
-                let value: Sub2Holder = map.next_value()?; // send to Self::visit_seq
-                return Ok(BaseAnyHolder::Sub2(Box::new(value)));
-            }
-            _ => {
-                use ::serde::de::{Error, Unexpected};
-                return Err(A::Error::invalid_value(Unexpected::Other(&key), &self));
-            }
-        }
-    }
-}
-
-impl WithVisitor for BaseAnyHolder {
-    type Visitor = BaseAnyHolderVisitor;
-    fn visitor_new() -> Self::Visitor {
-        BaseAnyHolderVisitor {}
-    }
 }
 
 #[test]
